my_pydantic_project/requirements.txt
annotated-types==0.7.0
ndiff==7.96
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
python-dotenv==1.2.2
six==1.17.0
typing-inspection==0.4.2
typing_extensions==4.15.0
wheel==0.46.3
